Add setup and rename to smt_rmo
