+++ setup.cfg
@@ -1,0 +1,4 @@
+[metadata]
+name = smt_rmo
+description-file = README.md
+
