+++ src/smt_rmo/rewriters.py
