+++ src/smt_rmo/strategies.py
@@ -10,8 +10,8 @@ from pysmt.fnode import FNode
 from pysmt.oracles import SizeOracle
 from pysmt.walkers import IdentityDagWalker
 
-from rmo.rewriters import Rewriter
-from rmo.utils import InfiniteTimeoutBudget, TimeoutBudget, track_timeout_budget
+from smt_rmo.rewriters import Rewriter
+from smt_rmo.utils import InfiniteTimeoutBudget, TimeoutBudget, track_timeout_budget
 
 
 class Visitor(ABC):
